Allows on-call queries to be case-insensitive
Fixes #40

Prior to this PR, asking an on-call question, such as `hubot who is on call for devops` would only produce a result if the name of the PagerDuty schedule was exactly "devops", but not if it were "DevOps". This PR converts the query text and result set item to lowercase, allowing a more permissive match.