Fix who's on call output
Found that running `who's on call` will give some garbage output:

```
hubot> hubot who's on call
4
6
5
3
1
2
7
8
* Foo Bar is on call for Example - Engineering - https://foo.pagerduty.com/schedules#PSH77FE
* Baz Biz is on call for Example - DEV - https://foo.pagerduty.com/schedules#chedules#PQ4T39J
```

Traced this back to ef1d84bd where the output was changed to return in a
single message rather several. Both the results of `messages.push()`
which is the length of the array and the actual `messages` array itself
was being sent back.