pagerduty: hide non-pageable schedules.
`.who's on call`'s primary use-case is people trying to figure out who
is on-call that should be paged for a schedule. As a result, `hubot` or
a non-pageable schedule are noise in this case (but useful for every
other command where e.g. users may wish to know if they are on-call for
a non-pageable shift).

As a result, don't display a schedule if `hubot` is on call or the
schedule has the pager disabled.