Render a better on-call list with bullets and links
We can take advantage of Slack's message formatting to make this output
easier to read. Specifically, we can remove the visual noise of the expanded
URLs (and instead linkifiy the name of the schedule), and we can also
easily include a link to the name of the person on-call (which may be useful
when trying to contact that person).