Chunk pagerduty schedule responses
Slack truncates messages around the 8000 character boundary, so we
should take care never to send a message longer than that length.

To do this, we add a chunking function, and send each chunk as a separate
message.

We could alternatively get around this by creating snippets or by using
"Posts" (which I think is relatively new), but this preserves the behavior
of the current integration a little bit better in my opinion.